+++ inc/invoke_cmd/redirect.h
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   redirect.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dayano <dayano@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ttsubo <ttsubo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/28 15:20:03 by dayano            #+#    #+#             */
-/*   Updated: 2025/04/28 15:30:35 by dayano           ###   ########.fr       */
+/*   Updated: 2025/05/04 19:19:04 by ttsubo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef REDIRECT_H
 # define REDIRECT_H
 
-# include "cmd.h"
+# include "common.h"
 
 void	redirect(t_cmd *cmd);
 void	redirect_stdout(t_cmd *cmd);
